moved infra provider tests to separate folder, refactoring
