+++ cfme/tests/openstack/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for Openstack cloud and infra providers"""
