+++ cfme/tests/openstack/infrastructure/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for Openstack infrastructure provider"""
