+++ cfme/tests/openstack/infrastructure/test_credentials_quad.py
@@ -12,7 +12,6 @@ pytest_generate_tests = testgen.generate(testgen.provider_by_type,
 
 @pytest.mark.usefixtures("setup_provider_modscope")
 def test_credentials_quads(provider):
-    provider.load_details()
     navigate_to(InfraProvider, 'All')
     quad = Quadicon(provider.name, qtype='infra_prov')
     checked = str(quad.creds).split('-')[0]
